Restore JRE 7 compatibility
The previous commits introduced some code that would not work on
JRE 7, which are fixed here.
